// Node.js, typescript/Typescript/Classes/src/enum/enums.ts
export enum Sex {
    Male = "MALE",
    Female = "FEMALE",
    XGender = "X"
}

export enum Examination {
    Exam = "EXAM",
    Test = "TEST",

}

export const maxSalary = 15000;

export enum Role {
    Cleaner = 'CLEANER',
    Cashier = 'CASHIER',
    Security = 'SECURITY',
    Manager = 'MANAGER',
    Loader = 'LOADER',
}

export enum Group {
    AppliedMathAppliad = "PMP",
    AppliedMathAnalysis = "PMA",
    AppliedMathInformatic = "PMI",
    AppliedLinguistEnglish = "PLE",
    AppliedLinguistFranch = "PLF",
    AppliedLinguistGermar = "PLG"
}
export enum Degree {
    Professor = "PROFESSOR",
    Docent = "DOCENT",
    Assistant = "ASSIASTANT",
    Masters = "MASTER"
}